impl consistsOf for all timespans
