+++ stats/stats/src/charts/types/timespans/mod.rs
@@ -5,5 +5,6 @@ mod year;
 
 pub use chrono::NaiveDate;
 pub use day::DateValue;
+pub use month::Month;
 pub use week::{Week, WeekValue};
 pub use year::Year;
